// src/api/swagger/models/Attachments.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Attachments = {
    attachmentId?: string;
    fileType?: string;
    filePath?: string;
    fileSize?: number;
    key?: string;
    uploadedAt?: string;
    fileName?: string;
};